requirements.txt
cufflinks==0.17.3
matplotlib==3.4.3
numpy==1.20.3
opencv_python==4.5.5.64
pandas==1.3.4
pandas_profiling==3.2.0
Pillow==9.1.1
plotly==5.7.0
pytz==2021.3
requests==2.26.0
scikit_learn==1.1.0
seaborn==0.11.2
streamlit==1.9.0
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
streamlit_pandas_profiling==0.1.3
